Output object files Linux can load
Linux and bcc do not expect the same format for functions (for map
references) and map declarations.  Therefore, this commit doubles all
sections holding functions and map declarations.  Map declarations are
simply created with the format expected by the Linux kernel.  Duplicates
of functions for bcc are appended a __fake__ marker.  The bcc version of
duplicated functions is rewritten to refer to maps with the appropriate
name; the Linux version refers to map declarations directly, without using
bpf_pseudo_fd.

Because of the differences when referring to maps, all inlined functions
must be duplicated as well, in case they refer to maps. Duplicating macros
is not supported, which is the reason sofdsnoop.py is failing.

This approach is quite fragile and was only tested for tools in the smoke
tests, several of which don't work.  Tools that have generic function
names and choose the hook point on the Python side will not work because
we're unable to detect the BPF program type at compile-time and we need it
to generate an appropriate section name for Linux.  In addition, any tool
using BPF_TABLE directly will not work as we require a new argument.

Signed-off-by: Paul Chaignon <paul.chaignon@orange.com>