Generate object files for tool in smoke tests
Smoke tests are convenient to generate the object files for a large number
of tools as they already have all the appropriate arguments for each tool.
Nevertheless, not all tools are part of the smoke tests.

This commit patches the smoke tests to rename the object files after the
tested tool in /tmp.

The tests were executed on a Linux 5.1.  Only sofdsnoop.py is failing at
the moment.

Signed-off-by: Paul Chaignon <paul.chaignon@orange.com>